+++ mountpoint-s3-fs/tests/common/fuse.rs
@@ -53,6 +53,8 @@ pub trait TestClient: Send {
 
     fn get_object_size(&self, key: &str) -> Result<usize, Box<dyn std::error::Error>>;
 
+    fn get_object_content_type(&self, key: &str) -> Result<Option<String>, Box<dyn std::error::Error>>;
+
     fn restore_object(&self, key: &str, expedited: bool) -> Result<(), Box<dyn std::error::Error>>;
 
     fn is_object_restored(&self, key: &str) -> Result<bool, Box<dyn std::error::Error>>;
@@ -461,6 +463,11 @@ pub mod mock_session {
             Ok(head_object.size as usize)
         }
 
+        fn get_object_content_type(&self, _key: &str) -> Result<Option<String>, Box<dyn std::error::Error>> {
+            // MockClient's HeadObject does not expose Content-Type headers.
+            Ok(None)
+        }
+
         fn restore_object(&self, key: &str, _expedited: bool) -> Result<(), Box<dyn std::error::Error>> {
             let full_key = format!("{}{}", self.prefix, key);
             Ok(self.client.restore_object(&full_key)?)
@@ -734,6 +741,12 @@ pub mod s3_session {
             let full_key = format!("{}{}", self.prefix, key);
             let head_object = tokio_block_on(self.sdk_client.head_object().bucket(&self.bucket).key(&full_key).send())?;
             Ok(head_object.content_length().unwrap() as usize)
+        }
+
+        fn get_object_content_type(&self, key: &str) -> Result<Option<String>, Box<dyn std::error::Error>> {
+            let full_key = format!("{}{}", self.prefix, key);
+            let head_object = tokio_block_on(self.sdk_client.head_object().bucket(&self.bucket).key(&full_key).send())?;
+            Ok(head_object.content_type().map(|s| s.to_owned()))
         }
 
         // Schedule restoration of an object, do not wait until completion. Expidited restoration completes within 1-5 min for GLACIER and is not available for DEEP_ARCHIVE.